requirements.txt
-i https://pypi.python.org/simple
aiodns==1.1.1
aiohttp==3.3.1
async-hvac==0.6.1
async-timeout==3.0.0
attrs==18.2.0
cchardet==2.1.4
chardet==3.0.4
docopt==0.6.2
envparse==0.2.0
idna-ssl==1.1.0
idna==2.7
jedi==0.12.1
marshmallow==2.15.6
multidict==4.4.2
parso==0.3.1
prompt-toolkit==1.0.15
ptpython==0.41
pycares==2.3.0
pygments==2.2.0
six==1.11.0
wcwidth==0.1.7
webargs==4.1.0
yarl==1.2.6
